refactor: refactor nav store to meta store
