+++ content/stores/content.ts
@@ -1,9 +1,6 @@
 import { defineStore, queryContent, useNuxtApp, useRoute } from '#imports'
-import { computed, ref } from 'vue'
-import { normalizePath } from '../utils'
-import { queryItems, queryTags } from './api'
+import { ref } from 'vue'
 import type { LocationQuery } from 'vue-router'
-import type { MetaItem, MetaPost, TagGroup } from '../types'
 
 /**
  * Raw content store
@@ -31,39 +28,6 @@ export const useContentStore = defineStore('content', () => {
   // page object
   const page = ref<PageContent | null>(null)
 
-  // all content items (metadata only)
-  const items = ref<MetaItem[]>([])
-
-  // tag groups from json
-  const tagGroups = ref<TagGroup[]>([])
-
-  // flattened tag list
-  const tagList = computed(() => {
-    return tagGroups.value
-      .map(group => group.tags)
-      .flat()
-      .sort()
-  })
-
-  // ---------------------------------------------------------------------------------------------------------------------
-  // actions
-  // ---------------------------------------------------------------------------------------------------------------------
-
-  /**
-   * Get all folder and posts from the target path down
-   */
-  function getItems (path = '/'): MetaItem[] {
-    const normalizedPath = normalizePath(path)
-    return items.value.filter(item => item.path.startsWith(normalizedPath))
-  }
-
-  /**
-   * Get all posts from the target path down
-   */
-  function getPosts (path = '/'): MetaPost[] {
-    return getItems(path).filter(p => p.type === 'post')
-  }
-
   /**
    * Load a page from the content api
    */
@@ -84,18 +48,8 @@ export const useContentStore = defineStore('content', () => {
   // ---------------------------------------------------------------------------------------------------------------------
 
   async function initServer () {
-    // set initial path value
     path.value = route.path
     query.value = route.query
-
-    // load all data
-    const results = await Promise.all([
-      // Can't have more than one await @see https://www.youtube.com/watch?v=ofuKRZLtOdY
-      queryItems(), // process.env.NODE_ENV as any
-      queryTags(),
-    ])
-    items.value = results[0]
-    tagGroups.value = results[1]
   }
 
   function initClient () {
@@ -110,15 +64,6 @@ export const useContentStore = defineStore('content', () => {
     // route
     path,
     query,
-
-    // tags
-    tagGroups,
-    tagList,
-
-    // items
-    items,
-    getItems,
-    getPosts,
 
     // page
     page,
